Refactor EnvOverride to be shared in tests/utils, change FILE to fstream, remove exceptions from toml++
